requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
pydantic-settings==2.7.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
psycopg2-binary==2.9.10
email-validator==2.2.0
pytest==8.3.4
httpx==0.28.1

